[RENAME] - auth Package 이름 변경
auth -> authentication
